ref: minor refactoring

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\TaskResource;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class TaskController extends Controller
 {
@@ -15,6 +17,7 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
     /**
@@ -22,6 +25,9 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if (request()->user()->cannot('create', Task::class)) {
+            abort(403, 'This action is unauthorized.');
+        }
         $task = $request->user()->tasks()->create($request->validated());
         return TaskResource::make($task);
     }
@@ -31,6 +37,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
+
         return TaskResource::make($task);
     }
 
@@ -39,6 +47,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403, 'This action is unauthorized.');
+        }
+
         $task->update($request->validated());
         return TaskResource::make($task);
     }
@@ -48,6 +60,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if (request()->user()->cannot('delete', $task)) {
+            abort(403, 'This action is unauthorized.');
+        }
+
         $task->delete();
         return response()->noContent();
     }
